requirements.txt
docxtpl
PyGithub
qrcode[pil]